+++ package.json
@@ -14,7 +14,7 @@
     "lint:fix": "eslint --fix",
     "test": "npm run test:unit && npm run test:typescript",
     "test:unit": "c8 node --test",
-    "test:typescript": "tsd"
+    "test:typescript": "tstyche"
   },
   "repository": {
     "type": "git",
@@ -95,7 +95,7 @@
     "pug": "^3.0.2",
     "split2": "^4.2.0",
     "squirrelly": "^9.1.0",
-    "tsd": "^0.33.0",
+    "tstyche": "^7.0.0",
     "twig": "^3.0.0"
   },
   "publishConfig": {
